+++ src/interfaces/DragonchainClientInterfaces.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Dragonchain, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2020 Dragonchain, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -314,6 +314,20 @@ export interface BinanceInterchainNetwork {
  */
 export interface InterchainNetworkList {
   interchains: BitcoinInterchainNetwork[] | EthereumInterchainNetwork[] | BinanceInterchainNetwork[];
+}
+
+/**
+ * Transaction hash (or equivalent) of a published interchain transaction
+ * @example
+ * ```json
+ *
+ * {
+ *   "transaction": "0x37d1b1032288f7b434973c7840123bbf1acdb5ef5e40dda94257208ccf6c0c9f"
+ * }
+ * ```
+ */
+export interface PublishedInterchainTransaction {
+  transaction: string;
 }
 
 /**