+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.3.0
+
+- **Feature**
+  - Add option for deleting smart contract by transaction type
+  - Add support for publishing signed interchain transactions
+  - Export `DragonchainClient` type for external use
+- **Development:**
+  - Update development dependencies
+  - Remove codeowners
+
 ## 4.2.1
 
 - **Feature:**